updated restrictions on organizations

+++ src/controllers/organizations.js
@@ -6,6 +6,7 @@ module.exports = {
     removeOrganization
 }
 const Organizations = require('../database/helpers/organizations');
+const Users = require('../database/helpers/users');
 
 async function getOrganizations(req, res){
     try {
@@ -18,7 +19,14 @@ async function getOrganizations(req, res){
 
 async function getOrganization(req, res){
     try {
+        const user = await Users.getUserById(req.user.id);
+
+        if(user.role !== 'ADMINISTRATOR' && user.organization_id !== req.params.id){
+            return await res.status(403).json({ error: 'This organization is not yours!' });
+        }
+
         const organization = await Organizations.getOrganizationById(req.params.id);
+        
         if(organization === undefined || organization.deleted) return await res.status(404).json({ error: 'Organization not found!' });
         return await res.status(200).json(organization);
     } catch(error) {
@@ -28,6 +36,11 @@ async function getOrganization(req, res){
 
 async function updateOrganization(req, res){
     try {
+        const user = await Users.getUserById(req.user.id);
+
+        if(user.role !== 'ADMINISTRATOR' && user.organization_id !== req.params.id){
+            return await res.status(403).json({ error: 'This organization is not yours!' });
+        }
         const organizationData = {
             name,
             description,
@@ -45,6 +58,11 @@ async function updateOrganization(req, res){
 
 async function deleteOrganization(req, res){
     try {
+        const user = await Users.getUserById(req.user.id);
+
+        if(user.role !== 'ADMINISTRATOR' && user.organization_id !== req.params.id){
+            return await res.status(403).json({ error: 'This organization is not yours!' });
+        }
         const organization = await Organizations.getOrganizationById(req.params.id);
         if(organization === undefined || organization.deleted) return await res.status(404).json({ error: 'Organization not found!' });
         await Organizations.deleteOrganization(req.params.id);
